+++ HISTORY.rst
@@ -116,9 +116,15 @@ History
 
 * Added support for check-less data types such as soil moisture
 
-0.6.5 (?)
+0.6.5 (2024-08-09)
 ----------------------------------
 
 * Made quality_encoder automatically assign qc200 for check-less data
 * Fixed the missing data quality codes to fit with hilltop's funky qc system
-* ?
+* Added batch processing
+* Yaml now specifies destination file name
+
+0.6.6 (2024-08-27)
+----------------------------------
+
+* Adding support for infer frequency and missing record prototype